mv mercatore root resource to root package
