+++ src/adhocracy_mercator/adhocracy_mercator/resources/test_mercator.py
@@ -11,18 +11,6 @@ from mercator.tests.fixtures.fixturesMercatorProposals1 import create_proposal_b
 from mercator.tests.fixtures.fixturesMercatorProposals1 import update_proposal_batch
 
 
-def test_root_meta():
-    from adhocracy_core.resources.root import root_meta
-    from adhocracy_core.resources.root import \
-        create_initial_content_for_app_root
-    from .mercator import _create_initial_content
-    from .mercator import mercator_root_meta
-    assert _create_initial_content not in root_meta.after_creation
-    assert _create_initial_content in mercator_root_meta.after_creation
-    assert create_initial_content_for_app_root in \
-        mercator_root_meta.after_creation
-
-
 def test_mercator_proposal_meta():
     from .mercator import mercator_proposal_meta
     from .mercator import IMercatorProposal
@@ -77,20 +65,13 @@ def integration(config):
     config.include('adhocracy_core.catalog')
     config.include('adhocracy_core.sheets')
     config.include('adhocracy_core.graph')
-    config.include('adhocracy_core.rest')
-    config.include('adhocracy_core.authentication')
     config.include('adhocracy_core.resources.geo')
     config.include('adhocracy_core.resources.tag')
-    config.include('adhocracy_core.resources.comment')
-    config.include('adhocracy_core.resources.rate')
-    config.include('adhocracy_core.resources.principal')
     config.include('adhocracy_core.resources.pool')
     config.include('adhocracy_core.resources.asset')
     config.include('adhocracy_core.resources.item')
-    config.include('adhocracy_core.resources.sample_paragraph')
-    config.include('adhocracy_core.resources.sample_proposal')
-    config.include('adhocracy_core.resources.sample_section')
-    config.include('adhocracy_core.resources.external_resource')
+    config.include('adhocracy_core.resources.comment')
+    config.include('adhocracy_core.resources.rate')
     config.include('adhocracy_mercator.sheets.mercator')
     config.include('adhocracy_mercator.resources.mercator')
     config.include('adhocracy_mercator.resources.subscriber')
@@ -123,12 +104,7 @@ class TestIncludemeIntegration:
                                       )
         assert IMercatorProposalVersion.providedBy(res)
 
-    def test_create_mercator_root_with_initial_content(self, registry):
-        from adhocracy_core.resources.root import IRootPool
-        from adhocracy_core.resources.asset import IPoolWithAssets
-        inst = registry.content.create(IRootPool.__identifier__)
-        assert IRootPool.providedBy(inst)
-        assert IPoolWithAssets.providedBy(inst['mercator'])
+
 
 @fixture(scope='class')
 def app_anonymous(app_anonymous):
