+++ src/adhocracy_mercator/adhocracy_mercator/resources/test_root.py
@@ -1,0 +1,29 @@
+from pytest import fixture
+from pytest import mark
+
+
+@fixture
+def integration(config, pool_graph_catalog):
+    config.include('adhocracy_mercator')
+
+
+def test_root_meta():
+    from adhocracy_core.resources.root import root_meta
+    from adhocracy_core.resources.root import \
+        create_initial_content_for_app_root
+    from .root import _create_initial_content
+    from .root import mercator_root_meta
+    assert _create_initial_content not in root_meta.after_creation
+    assert _create_initial_content in mercator_root_meta.after_creation
+    assert create_initial_content_for_app_root in \
+        mercator_root_meta.after_creation
+
+
+@mark.usefixtures('integration')
+def test_create_root_with_initial_content(registry):
+        from adhocracy_core.resources.root import IRootPool
+        from adhocracy_core.resources.asset import IPoolWithAssets
+        inst = registry.content.create(IRootPool.__identifier__)
+        assert IRootPool.providedBy(inst)
+        assert IPoolWithAssets.providedBy(inst['mercator'])
+
